requirements.txt
streamlit>=1.36.0
pymupdf>=1.24.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
transformers>=4.41.2
torch>=2.2.0
numpy>=1.26.4
pandas>=2.2.2
openai>=1.40.0
tqdm>=4.66.4
